+++ CHANGELOG.md
@@ -5,6 +5,14 @@ file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## 26.0.0
+
+**This release adds support for Protocol 26**
+
+### Added
+- Added support for CAP-0077 (new result codes for protocol 26) ([#170](https://github.com/stellar/stellar-horizon/pull/170)).
+- Added CAP-0073 ingestion support, including new effects for minting, clawback, and transfer of regulated assets ([#171](https://github.com/stellar/stellar-horizon/pull/171)).
+
 ## 25.1.0
 
 ### Updated